app/requirements.txt
# Google ADK
google-adk>=1.10.0

# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Document processing
PyPDF2>=3.0.0
lxml>=4.9.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
psutil>=5.9.0
email-validator>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0 